requirements.txt
coverage==7.6.3
python-dotenv==1.0.1